fix: Document pages now hydrate correctly
- Extracted API resolvers, so they can be used by the client.
- Removed all CommonJS imports and exports (except for next.config.js).
- Fixed wrong set of video pages getting pre-rendered in staging mode.
- Fixed missing cover image on summary cards if title is too long.
- Changed target of index buttons on home page to the actual index pages.
- Added buttons to the home page for series, studio and page index.
- Moved index pages from /index/... to /anime, /artist, /series and /studio.
- Disabled prefetching for links on the index pages.
- Recently added themes are now fetched on the client, so they always show the most up-to-date data.
