--- src/lib/server/animethemes/enums.js
@@ -1,43 +1,0 @@
-exports.animeSeason = new Map([
-    [ 0, "Winter" ],
-    [ 1, "Spring" ],
-    [ 2, "Summer" ],
-    [ 3, "Fall" ]
-]);
-
-exports.videoSource = new Map([
-    [ 0, "WEB" ],
-    [ 1, "RAW" ],
-    [ 2, "BD" ],
-    [ 3, "DVD" ],
-    [ 4, "VHS" ],
-    [ 5, "LD" ]
-]);
-
-exports.videoOverlap = new Map([
-    [ 0, "None" ],
-    [ 1, "Transition" ],
-    [ 2, "Over" ]
-]);
-
-exports.imageFacet = new Map([
-    [ 0, "Small Cover" ],
-    [ 1, "Large Cover" ]
-]);
-
-exports.resourceSite = new Map([
-    [ 0, "Official Website" ],
-    [ 1, "Twitter" ],
-    [ 2, "aniDB" ],
-    [ 3, "AniList" ],
-    [ 4, "Anime-Planet" ],
-    [ 5, "Anime News Network" ],
-    [ 6, "Kitsu" ],
-    [ 7, "MyAnimeList" ],
-    [ 8, "Wiki" ]
-]);
-
-exports.themeType = new Map([
-    [ 0, "OP" ],
-    [ 1, "ED" ]
-]);
